Added life system
Added new life system mostly handled by the LoseScene. Some parts handled by the failPlayer function in Enemysprite. The player now can return to menu or continue back at the school entrance when they die if their grade is above a D. If they die and their grade is a D, then they lose and can either restart or return to the menu.